[Refactor] procedurecontent로 패키징
