+++ app/src/main/java/kr/sparta/tripmate/ui/budget/ProcedureDetailActivity.kt
@@ -9,6 +9,7 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import kr.sparta.tripmate.databinding.ActivityProcedureDetailBinding
+import kr.sparta.tripmate.ui.budget.procedurecontent.ProcedureContentActivity
 import kr.sparta.tripmate.ui.viewmodel.budget.ProcedureDetailViewModel
 import kr.sparta.tripmate.ui.viewmodel.budget.ProcedureDetailFactory
 import kr.sparta.tripmate.util.method.toMoneyFormat
