+++ app/src/main/java/kr/sparta/tripmate/ui/budget/budgetdetail/main/BudgetDetailActivity.kt
@@ -14,7 +14,7 @@ import kr.sparta.tripmate.R
 import kr.sparta.tripmate.data.model.budget.Budget
 import kr.sparta.tripmate.databinding.ActivityBudgetDetailBinding
 import kr.sparta.tripmate.ui.budget.budgetcontent.BudgetContentActivity
-import kr.sparta.tripmate.ui.budget.ProcedureContentActivity
+import kr.sparta.tripmate.ui.budget.procedurecontent.ProcedureContentActivity
 import kr.sparta.tripmate.ui.viewmodel.budget.budgetdetail.main.BudgetDetailFactory
 import kr.sparta.tripmate.ui.viewmodel.budget.budgetdetail.main.BudgetDetailViewModel
